testing

+++ src/testing/CuFFTExample.cu
@@ -18,7 +18,8 @@ int main(){
   std::vector<float> weights_RE { 1.0, 0.7, 0.5, 0.2, 0.3, 0.7, 0.8 };
   std::vector<float> weights_IM { 1.0, 0.3, 0.2, 0.4, 0.9, 0.1, 0.6 };
   std::unique_ptr<__half2[]> data =
-      CreateSineSuperpostionH2GPU(fft_length, weights_RE, weights_IM, 7);
+      CreateSineSuperpostionH2GPU(fft_length, weights_RE, weights_IM,
+                                  static_cast<long long>(7));
 
   __half2* dptr_data;
   __half2* dptr_results;
